Issue #23: Partial updates in response to PR reviews. ... Still in progress. ...
