+++ continuous-integration/starter-kits/compile-release-publish/java-maven/multimodule/multimodule-api/src/main/java/gov/nasa/jpl/ammos/ids/slim/sample_projects/multimodule/api/NumericViews.java
@@ -1,25 +1,17 @@
-package gov.nasa.ammos.ids.sample_projects.simple;
+/**
+ * Sample code -- Replace with your packages, classes and files
+ */
+package gov.nasa.ammos.ids.slim.sample_projects.multimodule.api;
 
 import java.lang.invoke.MethodHandles;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class HelloWorld {
+public class NumericViews {
 
 	private static Logger log = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass().getCanonicalName());
 
-	public static void main(String[] args) {
-
-		int count_start = 0;
-
-		log.info("Starting count. ...");
-		count_print(count_start);
-		log.info("Finished count. ...");
-		System.out.println("Hello World!");
-
-	}
-
-	static void count_print(int in_startpoint) {
+	public static void count_print(int in_startpoint) {
 		// in_startpoint defaults to 0
 		String out = "";
