+++ continuous-integration/starter-kits/compile-release-publish/java-maven/multimodule/multimodule-api/src/test/resources/simplelogger.properties
@@ -1,3 +1,7 @@
+##
+# Sample code -- Replace with your configurations
+##
+
 # SLF4J's SimpleLogger configuration file
 # Simple implementation of Logger that sends all enabled log messages, for all defined loggers, to System.err.
